# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KenyaPay Payroll System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leave/types:
    get:
      operationId: listLeaveTypes
      tags: [leave]
      summary: List all leave types
      responses:
        "200":
          description: Leave types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveType"
    post:
      operationId: createLeaveType
      tags: [leave]
      summary: Create a leave type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveTypeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveType"
  /leave/requests:
    get:
      operationId: listLeaveRequests
      tags: [leave]
      summary: List leave requests
      parameters:
        - in: query
          name: employeeId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leave]
      summary: Submit a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/requests/{id}:
    get:
      operationId: getLeaveRequest
      tags: [leave]
      summary: Get a leave request by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    patch:
      operationId: updateLeaveRequest
      tags: [leave]
      summary: Approve or reject a leave request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leave/balance/{employeeId}:
    get:
      operationId: getLeaveBalance
      tags: [leave]
      summary: Get leave balance for an employee
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveBalance"
...
components:
  schemas:
...
    LeaveType:
      type: object
      required:
        - id
        - name
        - daysAllowed
      properties:
        id:
          type: integer
        name:
          type: string
        daysAllowed:
          type: integer
        description:
          type: ["string", "null"]
    LeaveTypeInput:
      type: object
      required:
        - name
        - daysAllowed
      properties:
        name:
          type: string
        daysAllowed:
          type: integer
        description:
          type: string
    LeaveRequest:
      type: object
      required:
        - id
        - employeeId
        - leaveTypeId
        - startDate
        - endDate
        - days
        - status
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: integer
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        approverId:
          type: ["integer", "null"]
        approverName:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        appliedAt:
          type: string
    LeaveRequestInput:
      type: object
      required:
        - employeeId
        - leaveTypeId
        - startDate
        - endDate
      properties:
        employeeId:
          type: integer
        leaveTypeId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    LeaveRequestUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [approved, rejected]
        approverId:
          type: integer
        rejectionReason:
          type: string
    LeaveBalance:
      type: object
      properties:
        leaveTypeId:
          type: integer
        leaveTypeName:
          type: string
        daysAllowed:
          type: integer
        daysUsed:
          type: integer
        daysRemaining:
          type: integer
...